Fixed my Flying as Offensive STAB drawing
I wasn't told that it would be shrunken and I watermark all my artwork in case it's posted elsewhere by people outside of Smogon, and since the watermark I use is the smallest legible pixel font possible it was obscured - I uploaded a version that I shrunk myself with a properly sized watermark again. It's named what it was in the folder. Sorry for the trouble.